refactor(be): 정산 도메인에서 현재 부적합한 데이터 타입 변경
* 현재 설계상 데이터 값을 소수점으로 만들어서 저장하지는 않으므로 BigDecimal 대신 Long 사용
* long 보단 Long 을 통한 표현력 강화(조금의 메모리 공간과 오버헤드로 인한 성능차이는 미미)
* 배달원/상점의 정산 건 수는 21억을 넘을 일이 정말 없으므로

+++ backend/src/main/java/com/deliveranything/domain/settlement/dto/projection/SettlementSummaryProjection.java
@@ -1,13 +1,11 @@
 package com.deliveranything.domain.settlement.dto.projection;
 
-import java.math.BigDecimal;
-
 public record SettlementSummaryProjection(
     Long totalTransactionCount,
     Long weeklyTransactionCount,
-    BigDecimal weeklySettledAmount,
-    BigDecimal monthlySettledAmount,
-    BigDecimal totalSettledAmount
+    Long weeklySettledAmount,
+    Long monthlySettledAmount,
+    Long totalSettledAmount
 ) {
 
 }
